requirements.txt
rich>=13.0.0
pyyaml>=6.0
openai>=1.0.0
google-generativeai>=0.3.0
prompt-toolkit>=3.0.0
questionary>=2.0.0
requests>=2.31.0
